Add shields.io badges

+++ upload.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 pandoc --from=markdown --to=rst --output=README.rst README.md
-rm -rf dist/
-python setup.py sdist bdist_wheel
-twine upload dist/*
+if [ $# -eq 1 ]; then
+    rm -rf dist/
+    python setup.py sdist bdist_wheel
+    twine upload dist/*
+fi
